refactor(stencil configuration): update stencil config files to prevent package name inconsistency
Different parts of the stencil config file use the package name to scope files or otherwise organize
output. The package name was a literal in all these parts, which could result in variations cropping
up when a developer changes the package name but forgets to update everywhere. To avoid this
scenario, the package name has been centrally defined as a constant and this constant has been used
everywhere in the stencil config file.