fix(vuepress): fix Vuepress live preview plugin configuration based on latest version
With the shift to the monorepo setup, this repository now runs on Node 14.x (due to the need for npm
7+). Switching to this version of Node JS has resulted in some packages (mainly vuepress-plugin-live
and it's internal dependency vue-live) getting upgraded to newer versions. Currently, this
repository's Vuepress implementation includes a custom template (www/.vuepress/previewLayout.vue)
that over-rides vue-live's default template. This over-ride template was based on vue-live 0.3.0.
The updated version of vue-live (1.3.0) uses a very different method to render a live code editor
and hence it's template differs slightly - this difference in the template caused a
visual/functional bug in the live-preview parts of the website's UI. This patch updates the
over-ride template in accordance with the default template of vue-live 1.3.0 to fix this issue.