feat(custom-react-wrapper): use Custom React Wrapper instead of Stencil's react wrapper
Stencil React wrapper does not have auto-import and register custom elements for dependent
components. Stencil released this feature in v2.9 but this is not yet implemented in React Wrapper.
This is a custom react wrapper based on Stencil's react wrapper which implements auto-import and
register custom elements for dependent components. This also enables tree-shaking.